chore: remove deprecated OAuth and MongoDB templates, update package.json scripts and dependencies
