+++ .eslintrc.js
@@ -1,0 +1,10 @@
+module.exports = {
+    parser: '@typescript-eslint/parser',
+    parserOptions: { ecmaVersion: 2020, sourceType: 'module' },
+    plugins: ['@typescript-eslint'],
+    extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+    rules: {
+      // your overrides
+    }
+  };
+
