+++ src/packages/github-oauth/templates/config/githubStrategy.ejs
