+++ src/packages/github-oauth/templates/routes/githubRoutes.ejs
