+++ src/packages/google-oauth/templates/config/googleStrategy.ejs
