+++ src/packages/google-oauth/templates/routes/googleRoutes.ejs
