+++ src/packages/google-oauth/utils/ensureAppJsHasOAuthSetup.ts
