+++ src/packages/node-mongodb-template/templates/Instructions.ejs
