+++ src/packages/node-mongodb-template/templates/app.ejs
