+++ src/packages/node-mongodb-template/templates/controllers/userController.ejs
