+++ src/packages/node-mongodb-template/templates/env.example.ejs
