+++ src/packages/node-mongodb-template/templates/models/user.ejs
