+++ src/packages/node-mongodb-template/templates/package.json.ejs
