+++ src/packages/node-mongodb-template/templates/routes/userRoutes.ejs
