+++ src/packages/node-postgress-template/templates/package.json.ejs
