+++ src/packages/node-postgress-template/templates/routes/userRoutes.ejs
