+++ src/packages/node-postgress-template/templates/schema.prisma.ejs
