+++ src/packages/resend-otp/templates/otp.ejs
