+++ src/packages/resend-otp/templates/otpFunctions.ejs
