+++ src/packages/resend-otp/templates/otpRoutes.ejs
