+++ src/utils/injectEnvVars.ts
