+++ tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "allowJs": true, // Temporary allow JavaScript files to be included in the project
+    "compilerOptions": {
+      "target": "ES2020",
+      "module": "ESNext",
+      "moduleResolution": "node",
+      "lib": ["ES2020"],
+      "outDir": "dist",
+      "rootDir": "src",
+      "strict": true,
+      "esModuleInterop": true,
+      "allowSyntheticDefaultImports": true,
+      "skipLibCheck": true,
+      "forceConsistentCasingInFileNames": true,
+      "resolveJsonModule": true,
+      "declaration": true,
+      "sourceMap": true,
+      "allowImportingTsExtensions": true,
+      "noEmit": true
+    },
+    "include": ["src/**/*", "bin/**/*"],
+    "exclude": ["node_modules", "dist"]
+  }
+
